Add selectable menu items
d2l-menu-item-radio and d2l-menu-item-checkbox are both selectable menu items, based on the underlying d2l-menu-item-behavior-selectable. This type of menu item has a boolean "selected" state. The radio implementation will change all the item's parent's children to be selected=false on selection, whereas the checkbox can be toggled on or off.